% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula_indice_duncan.R
\name{calcula_indice_duncan}
\alias{calcula_indice_duncan}
\title{Cálculo dos Índices de Duncan}
\usage{
calcula_indice_duncan(dataset_selecionado, nivel_geografico, referencia, grupo)
}
\arguments{
\item{dataset_selecionado}{Dataset com colunas da tabela `dados_preprocessados` selecionadas.}

\item{nivel_geografico}{Nível geográfico para o cálcul do índice. Valores podem ser: `MN` - para município; `RM` para região metropolitana; `UF` para Unidade Federativa; `RG` para Região e `Brasil` para o país.}

\item{referencia}{Variável que será usada como referência para o cálulo do Índice. Equivale ao `X` da fórmula.}

\item{grupo}{Variável que contém o grupo a ser usado como cálculo do Índice. Equivale ao `Y` da fórmula.}
}
\value{
Uma tabela contendo os resultados do índice para o dado nível geográfico
}
\description{
Cálculo dos Índices de Duncan
}
