% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constroi_agregacao_variaveis.R
\name{constroi_agregacao_variaveis}
\alias{constroi_agregacao_variaveis}
\title{Constroi Novas Colunas por Agregação}
\usage{
constroi_agregacao_variaveis(
  dataset_selecionado,
  nova_variavel,
  variaveis_agregadas
)
}
\arguments{
\item{dataset_selecionado}{Dataset com colunas da tabela `dados_preprocessados` selecionadas.}

\item{nova_variavel}{Nome da nova variável a ser criadas.}

\item{variaveis_agregadas}{Nome das colunas que serão somadas para criar a nova coluna}
}
\value{
Uma tabela com as novas variaveis criadas segundo a agregação especificada.
}
\description{
Essa função é responsável por criar novas colunas via agregação de outras colunas. Por exemplo "Não Brancas" é o resultado da da soma do resultado para Raça ou Cor Preta e Parda.
}
