% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constroi_calculo_indice_estrutura.R
\name{constroi_calculo_indice_estrutura}
\alias{constroi_calculo_indice_estrutura}
\title{Construção da Estrutura do Cálculo dos Índices}
\usage{
constroi_calculo_indice_estrutura(dataset_selecionado)
}
\arguments{
\item{dataset_selecionado}{Dataset com colunas selecionadas da tabela `dados_preprocessados`.}
}
\value{
Uma tabela contendo os níves geográficos, variáveis de referência e variáveis contendo os grupos em que os índices serão baseados.
}
\description{
Essa função é responsável por construi
}
