% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constroi_sumarizacoes.R
\name{constroi_sumarizacoes}
\alias{constroi_sumarizacoes}
\title{Constrói Sumarização das Variáveis}
\usage{
constroi_sumarizacoes(dataset_selecionado, nivel_geografico)
}
\arguments{
\item{dataset_selecionado}{Dataset com colunas da tabela `dados_preprocessados` selecionadas.}

\item{nivel_geografico}{Nível geográfico para o cálcul do índice. Valores podem ser: `MN` - para município; `RM` para região metropolitana; `UF` para Unidade Federativa; `RG` para Região e `Brasil` para o país.}
}
\value{
Uma tabela contendo os dados sumarizados.
}
\description{
Constrói Sumarização das Variáveis
}
