% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputa_kernel.R
\name{imputa_kernel}
\alias{imputa_kernel}
\title{Imputação do Modelo de Kernel}
\usage{
imputa_kernel(dataset_cleaned)
}
\arguments{
\item{dataset_cleaned}{}
}
\value{
Uma tabela com as colunas tratadas adicionadas. As novas colunas contém
o sufixo indetificador `knn` para aplicação do modelo de KNN para o tratamento de 
dados faltantes e `kernel` para aplicação do modelo de Kernel.
}
\description{
Essa função é responsável para por imputar um modelo de KNN e Kernel aos dados
faltantes.
}
