% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessar_dados_brutos.R
\name{preprocessar_dados_brutos}
\alias{preprocessar_dados_brutos}
\title{Preprocessamento dos Dados Brutos}
\usage{
preprocessar_dados_brutos(
  dados_agregados_setor_censitario = dpupsetl::dat_agregados_setor_censitario,
  dicionario_agregados_setor_censitario = dpupsetl::dic_agregados_setor_censitario,
  dados_regioes_metropolitanas = dpupsetl::dat_regioes_metropolitanas
)
}
\arguments{
\item{dados_agregados_setor_censitario}{Tabela contendo os dados Agregados do Setor Censitário.}

\item{dicionario_agregados_setor_censitario}{Tabela contendo o Dicionário dos dados de Agregados do Setor Censitário.}

\item{dados_regioes_metropolitanas}{Tabela contendo os dados das Regiões Metropolitanas do Brasil.}
}
\value{
Salva ma tabela no formato RDS contendo os dados de Agregados do Setor Censitário tratados.
Esses dados estão pronto para serem inseridos na função `calcula_indice_duncan`.
}
\description{
Essa função é responsável por aplicar um tratamento de limpeza e dados
}
