% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforma_dados.R
\name{transforma_dados}
\alias{transforma_dados}
\title{Transformação dos Dados}
\usage{
transforma_dados(
  modelo = "kernel",
  dados_agregados_setor_censitario = dpupsetl::dat_agregados_setor_censitario,
  dicionario_agregados_setor_censitario = dpupsetl::dic_agregados_setor_censitario,
  dados_regioes_metropolitanas = dpupsetl::dat_regioes_metropolitanas,
  usar_dados_preprocessados = T
)
}
\arguments{
\item{modelo}{Colunas que serão utilizadas com base no uso de um modelo para o tratamento de daos faltantes. Valores podem ser `knn`, `kernel` ou `NULL` para coluna sem tratamento.}

\item{dados_agregados_setor_censitario}{Tabela contendo os dados Agregados do Setor Censitário.}

\item{dicionario_agregados_setor_censitario}{Tabela contendo o Dicionário dos dados de Agregados do Setor Censitário.}

\item{dados_regioes_metropolitanas}{Tabela contendo os dados das Regiões Metropolitanas do Brasil.}
}
\value{
Salva as tabelas:
\itemize{
\item{dados_tabela}{Tabela salva em formato .rds na pasta `inst/extdata/` que contém os dados que são consumidos pelo Painel.}
\item{dados_completos}{Tabela salva em format .csv na pasta `inst/extdata` que contém os dados completos que são baixados no Painel. Essa tabela deve ser atualizada no drive.}
}
}
\description{
Função responsável por aplicar o processo de transformação nos dados, construindo os dados de Índice de Duncan e resultados das variáveis por níveis geográficos.
}
